fix save_model_to_string for large model (#1080)
* use int64 for string

* [R] Fatal when CSC exceed int32.max

+++ src/c_api.cpp
@@ -1140,13 +1140,13 @@ int LGBM_BoosterSaveModel(BoosterHandle handle,
 #pragma warning(disable : 4996)
 int LGBM_BoosterSaveModelToString(BoosterHandle handle,
                                   int num_iteration,
-                                  int buffer_len,
-                                  int* out_len,
+                                  int64_t buffer_len, 
+                                  int64_t* out_len,
                                   char* out_str) {
   API_BEGIN();
   Booster* ref_booster = reinterpret_cast<Booster*>(handle);
   std::string model = ref_booster->SaveModelToString(num_iteration);
-  *out_len = static_cast<int>(model.size()) + 1;
+  *out_len = static_cast<int64_t>(model.size()) + 1;
   if (*out_len <= buffer_len) {
     std::strcpy(out_str, model.c_str());
   }
@@ -1156,13 +1156,13 @@ int LGBM_BoosterSaveModelToString(BoosterHandle handle,
 #pragma warning(disable : 4996)
 int LGBM_BoosterDumpModel(BoosterHandle handle,
                           int num_iteration,
-                          int buffer_len,
-                          int* out_len,
+                          int64_t buffer_len,
+                          int64_t* out_len,
                           char* out_str) {
   API_BEGIN();
   Booster* ref_booster = reinterpret_cast<Booster*>(handle);
   std::string model = ref_booster->DumpModel(num_iteration);
-  *out_len = static_cast<int>(model.size()) + 1;
+  *out_len = static_cast<int64_t>(model.size()) + 1;
   if (*out_len <= buffer_len) {
     std::strcpy(out_str, model.c_str());
   }
